requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.8
decorator==5.1.1
easydict==1.13
easyocr==1.7.2
einops==0.8.0
executing==2.1.0
fastapi==0.115.8
fastapi-cli==0.0.5
#ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.2.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httptools==0.6.4
humanfriendly==10.0
idna==3.10
imageio==2.35.1
insightface==0.7.3
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.13.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.3
scenedetect==0.6.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==70.0.0
shapely==2.0.6
shellingham==1.5.4
simsimd==6.2.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.3.0+cpu
torchaudio==2.3.0+cpu
torchvision==0.18.1+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.92
ultralytics-thop==2.0.14
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
exceptiongroup==1.2.2
python-box==7.3.2